Rename shaders.ts to shaderUtils.ts.
